requirements/base.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pytz==2024.2
sqlparse==0.5.1
psycopg2
python-dotenv==1.0.1
black==24.8.0
drf-yasg==1.21.7
setuptools==75.1.0
isort==5.13.2
flake8==7.1.1
django-cors-headers==3.14.0
